Required changes to support modified GetBestMove
GetBestMove was modified because I couldn't think of a cleaner way of doing things. If you know a better way (there has to be) then I'd say go ahead and implement it. I would've done a struct, but I wasn't sure if that would be okay with the project or not.