Exception during lc0 startup - using backend opts (#421)
* Exception during lc0 startup - using backend opts

When backend options are used like: 
lc0.exe selfplay --backend-opts="cudnn(threads=1, minibatch-size=512)"
the variable dict used before will get invalid, when the methode FromString will end, which lead to an exception in the debug mode and might lead to bigger issues later in the runtime mode, when the options are overwritten by random.

* Update optionsdict.cc

When backend options are used like: 
lc0.exe selfplay --backend-opts="cudnn(threads=1, minibatch-size=512)"
the variable dict used before will get invalid, when the methode FromString will end, which lead to an exception in the debug mode and might lead to bigger issues later in the runtime mode, when the options are overwritten by random.

* Update optionsdict.h

Remove the static member function from string and use a standard member function. Add a comment to avoid that this object will go out of scope, but is still referenced in the parent pointer.

* update optionsdict.cc

Move FromString to a member function.

* Update tournament.cc

Call the new member function AddSubdictFromString.

* Update engine.cc

Call the new member function AddSubdictFromString

* Update file

Solve a merge conflict. I'm sorry. Not 100% sure how to update a pull request