Create readme_for_dummies
Some people would probably need someone to speak to them a translation of the technical babble and a verbatim description of the instructions and troubleshooting for certain OS's may reduce people missing any steps. This file is far from complete, but it is a start for macOS users at the least.