Update README.md
It doesn't make any sense to have to install clang AND g++ on Ubuntu 18.04 when the build instructions explicitly say you're building with clang and the environment variables are enforcing that to the meson call. If you're building with clang, you don't need g++ at all. You DO need one of g++-8.0's dependencies, which is libstdc++-8. So this fix to the documentation makes that more clear and keeps the user from unnecessarily installing a whole compiler when just a library that the compiler depends on will do.