Fix one critical bug and one tiny bug.
The first item in each run after an erase/swap_to_erased was incorrectly *always* considered to be InRange - so unless the run continued past that point, it would not be swapped in to the correct position, and nor would anything else. Thus the item could no longer be found in the cache.
Luckily Evict code is overzealous and will always remove the item eventually (at a large cpu cost) to avoid cache flood, but inability to find the item in the cache normally means unpin doesn't work, so such items all go into an ever growing evict list, which progressively makes unpin slower and slower.
also was missing a return path when removing a pin from evict list that wasn't the last pin. Meaning it would pointlessly search the main list, fail and then trigger the debug assert (if debug asserts were enabled in the build...)