Enable SYCL NVIDIA and AMD backends (#2192)
* Enable SYCL NVIDIA backend

Tested with
```
CXX=icpx CC=icx AR=llvm-ar ./build.sh -Dsycl=nvidia
```
on Ubuntu 24.04 with CUDA 12.9 and oneAPI 2025.1.

The CUDA Compute Capability can be optionally specified with `-Dcc_cuda`.
If not specified, the default CUDA target of the DPC++ compiler is used,
which means SYCL device code is precompiled for the lowest supported CC.
When executed on a GPU with different CC, it is recompiled at runtime
for the specific architecture.

In addition to meson.build changes, remove a redundand free(nullptr)
causing crashes in the SYCL NVIDIA backend.

* Enable SYCL AMD backend and fix its fp16 support

Tested with
```
CXX=icpx CC=icx AR=llvm-ar ./build.sh -Dsycl=amd -Damd_gfx=90a
```
on Ubuntu 22.04 with ROCm 6.3.3 and oneAPI 2025.1.

The new amd_gfx option is required as DPC++ does not support
Just-In-Time compilation for AMD GPU code. It has to be precompiled
for the right architecture when building the application.

Fix the SYCL AMD fp16 backend which missed calling the fp16 hipBLAS
functions where needed.

Also fix the hardcoded sub-group / warp / wavefront size of 32. Some
AMD GPUs have wavefront size of 64 and this has to be used instead.

* Move SYCL_SUB_GROUP_SIZE definition to a common header