Simplify task reading code
Atomic variables can be used without locks. We want an id from the
shared taksks_taken_ counter. We want to increment the counter while
staying less than or equal to the task_count_. This can be done using
compare and exchange operation. We call compare exchange in a loop
because it will either success soon or other thread increment counters
to an equal value.