Calculate edge score only once
It appears that aggressive compiler optimisation can cause compare
function calculate different floating point values. This results to
undefined behavior. This is a potential compiler bug because only
-ffast-math should allow dangerous math optimisations.

We can compute the score once for shorting to avoid different values for
comparison. This becomes a key and value sorting problem which is
natural to do using std::multimap.